refactor(tenants): extract AvatarTab component
